Terraform/function-face-cut/requirements.txt
requests
boto3
Pillow
ydb
